protyping orion verify without merkle tree querying
